Move errors to general file module
