+++ src/biotite/structure/io/pdbx/cif.py
@@ -11,8 +11,8 @@ import shlex
 from collections.abc import MutableMapping, Sequence
 import numpy as np
 from .component import _Component, MaskValue
-from .error import DeserializationError, SerializationError
-from ....file import File, is_open_compatible, is_text
+from ....file import File, is_open_compatible, is_text, DeserializationError, \
+                     SerializationError
 
 
 UNICODE_CHAR_SIZE = 4
